+++ packages/server/src/api/admin/index.ts
@@ -1,14 +1,20 @@
 import {Hono} from 'hono';
-import {ServerOptions} from '../../types.js';
+import {Assert, IsZodExactly, ServerOptions} from '../../types.js';
 import {setup} from '../../setup.js';
-import {typiaValidator} from '@hono/typia-validator';
-import {createValidate} from 'typia';
 import {Env} from '../../env.js';
+import {z} from 'zod';
+import {zValidator} from '@hono/zod-validator';
+
+type ResetAction = {reset: boolean};
+const ResetActionSchema = z.object({
+	reset: z.boolean(),
+});
+type ZodMatchResetAction = Assert<IsZodExactly<typeof ResetActionSchema, ResetAction>>;
 
 export function getAdminAPI<Bindings extends Env>(options: ServerOptions<Bindings>) {
 	const app = new Hono<{Bindings: Bindings}>()
 		.use(setup({serverOptions: options}))
-		.post('/db-reset', typiaValidator('json', createValidate<{reset: boolean}>()), async (c) => {
+		.post('/db-reset', zValidator('json', ResetActionSchema), async (c) => {
 			const config = c.get('config');
 			const env = config.env;
 			const storage = config.storage;