+++ packages/server/src/api/private/index.ts
@@ -1,24 +1,24 @@
 import {Hono} from 'hono';
-import {ServerOptions} from '../../types.js';
-import {getAuth, setup} from '../../setup.js';
-import {typiaValidator} from '@hono/typia-validator';
-import {createValidate} from 'typia';
 import {
-	ActionAcceptConversation,
-	ActionGetAcceptedConversations,
-	ActionGetConversations,
-	ActionGetMessages,
-	ActionGetUnacceptedConversations,
-	ActionMarkAsRead,
-	ActionSendMessage,
-	ActionRejectConversation,
-} from 'missiv-common';
+	ActionAcceptConversationSchema,
+	ActionGetAcceptedConversationsSchema,
+	ActionGetConversationsSchema,
+	ActionGetMessagesSchema,
+	ActionGetUnacceptedConversationsSchema,
+	ActionMarkAsReadSchema,
+	ActionRejectConversationSchema,
+	ActionSendMessageSchema,
+	ServerOptions,
+} from '../../types.js';
+import {getAuth, setup} from '../../setup.js';
+import {zValidator} from '@hono/zod-validator';
+
 import {Env} from '../../env.js';
 
 export function getPrivateChatAPI<Bindings extends Env>(options: ServerOptions<Bindings>) {
 	const app = new Hono<{Bindings: Bindings}>()
 		.use(setup({serverOptions: options}))
-		.post('/sendMessage', typiaValidator('json', createValidate<ActionSendMessage>()), async (c) => {
+		.post('/sendMessage', zValidator('json', ActionSendMessageSchema), async (c) => {
 			const config = c.get('config');
 			const storage = config.storage;
 			const timestampMS = Date.now();
@@ -35,7 +35,7 @@ export function getPrivateChatAPI<Bindings extends Env>(options: ServerOptions<B
 			const result = await storage.sendMessage(publicKey, account, timestampMS, action);
 			return c.json(result);
 		})
-		.post('/getConversations', typiaValidator('json', createValidate<ActionGetConversations>()), async (c) => {
+		.post('/getConversations', zValidator('json', ActionGetConversationsSchema), async (c) => {
 			const config = c.get('config');
 			const storage = config.storage;
 			const {account} = getAuth(c);
@@ -48,43 +48,35 @@ export function getPrivateChatAPI<Bindings extends Env>(options: ServerOptions<B
 			const result = await storage.getConversations(action.domain, action.namespace, account);
 			return c.json(result);
 		})
-		.post(
-			'/getAcceptedConversations',
-			typiaValidator('json', createValidate<ActionGetAcceptedConversations>()),
-			async (c) => {
-				const config = c.get('config');
-				const storage = config.storage;
+		.post('/getAcceptedConversations', zValidator('json', ActionGetAcceptedConversationsSchema), async (c) => {
+			const config = c.get('config');
+			const storage = config.storage;
 
-				const {account} = getAuth(c);
-				if (!account) {
-					throw new Error(`no account authenticated`);
-				}
+			const {account} = getAuth(c);
+			if (!account) {
+				throw new Error(`no account authenticated`);
+			}
 
-				const action = c.req.valid('json');
+			const action = c.req.valid('json');
 
-				const result = await storage.getAcceptedConversations(action.domain, action.namespace, account);
-				return c.json(result);
-			},
-		)
-		.post(
-			'/getUnacceptedConversations',
-			typiaValidator('json', createValidate<ActionGetUnacceptedConversations>()),
-			async (c) => {
-				const config = c.get('config');
-				const storage = config.storage;
+			const result = await storage.getAcceptedConversations(action.domain, action.namespace, account);
+			return c.json(result);
+		})
+		.post('/getUnacceptedConversations', zValidator('json', ActionGetUnacceptedConversationsSchema), async (c) => {
+			const config = c.get('config');
+			const storage = config.storage;
 
-				const {account} = getAuth(c);
-				if (!account) {
-					throw new Error(`no account authenticated`);
-				}
+			const {account} = getAuth(c);
+			if (!account) {
+				throw new Error(`no account authenticated`);
+			}
 
-				const action = c.req.valid('json');
+			const action = c.req.valid('json');
 
-				const result = await storage.getUnacceptedConversations(action.domain, action.namespace, account);
-				return c.json(result);
-			},
-		)
-		.post('/markAsRead', typiaValidator('json', createValidate<ActionMarkAsRead>()), async (c) => {
+			const result = await storage.getUnacceptedConversations(action.domain, action.namespace, account);
+			return c.json(result);
+		})
+		.post('/markAsRead', zValidator('json', ActionMarkAsReadSchema), async (c) => {
 			const config = c.get('config');
 			const storage = config.storage;
 
@@ -99,7 +91,7 @@ export function getPrivateChatAPI<Bindings extends Env>(options: ServerOptions<B
 			return c.json({success: true});
 		})
 
-		.post('/rejectConversation', typiaValidator('json', createValidate<ActionRejectConversation>()), async (c) => {
+		.post('/rejectConversation', zValidator('json', ActionRejectConversationSchema), async (c) => {
 			const config = c.get('config');
 			const storage = config.storage;
 
@@ -113,7 +105,7 @@ export function getPrivateChatAPI<Bindings extends Env>(options: ServerOptions<B
 			await storage.rejectConversation(account, action);
 			return c.json({success: true});
 		})
-		.post('/getMessages', typiaValidator('json', createValidate<ActionGetMessages>()), async (c) => {
+		.post('/getMessages', zValidator('json', ActionGetMessagesSchema), async (c) => {
 			const config = c.get('config');
 			const storage = config.storage;
 
@@ -127,7 +119,7 @@ export function getPrivateChatAPI<Bindings extends Env>(options: ServerOptions<B
 			const result = await storage.getMessages(account, action);
 			return c.json(result);
 		})
-		.post('/acceptConversation', typiaValidator('json', createValidate<ActionAcceptConversation>()), async (c) => {
+		.post('/acceptConversation', zValidator('json', ActionAcceptConversationSchema), async (c) => {
 			const config = c.get('config');
 			const storage = config.storage;