+++ packages/server/src/schema/sql/02_user_alter.sql
@@ -1,1 +1,0 @@
-ALTER TABLE DomainUsers ADD COLUMN settings text;